[build] fix build
ATM we need node option "-max-old-space-size=8192" to be set for the
build to succeed, when one runs "yarn" on the repo.

This was already done in package.json, but it looks-like the option
is not not being passed-down to sub-scripts, and so the build fails.

I played with this a bit and found how that if I pass the option
directly for scripts "compile:extensions" and "bundle:extensions",
then it works.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>