[examples] Update dependencies, README
ATM yarn.lock pins theia extensions to a really old version. I have removed
it and generated a new one that pulls up-to-date Theia extensions and other
dependencies.

As well the Electron example app was not building/working, I think because
of upstream changes between the previously pinned Theia extensions version
and the new one. I have added @theia/electron as dev-dependency to the
Electron example application to fix this.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>