update vsx publish script
The conclusion of recent discussions is that the only differences
between "latest" and "next" extensions we publish to open-vsx is
the `preview` flag in the extension's package.json and the version
we'll use.

Both these things are taken-care of during packaging, so the publish
script no longer needs to use an argument; it will publish whatever
has been packaged.
---

recreated yarn.lock

This permits using a recent theia for the example apps. While there
Added @theia/vsx-registry in the examples. I needed to add a resolution
for the build to still succesed, after adding the new extension.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>