[ci] add GH action to publish specific, solid version
It can happen that we wish to re-publish an older version of the built-ins.
ATM it's not so easy. This PR adds an action that attempts to make it easier:

Push to branch `vsx-publish` to trigger a build and publish of the vscode-builtins.
Make sure the vscode git submodule is checked-out to the targetted version.

Note: to republish, the version needs to be removed from the registry first.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>