add GH action to publish a specific next version
This commit adds a GH action to publish an arbitrary next version of
the vscode-builtins: `publish-vsx-specific-next`. For this to work,
the version to be published needs to be removed from open-vsx, if
already present. Then one must push to branch `ovsx-publish-next`,
whatever they want built. e.g. the most recent master version of the
repo, but with a specific commit checked in the vscode git submodule.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>