Add script to generate a built-in extension pack
The new script 'create-extension-pack.js' can be executed directly
after a project build (i.e. yarn) as long as the extensions
associated with the build are already available in the registry.

When the extensions are not yet available in the registry the script
'package-vsix.js' needs to be executed as a previous step.

The result will be an extension package in .vsix format installed
under the 'dist' folder (which is then used as the source for the
existing publishing step).

This extension package will include a reference to all extensions
resolved during the build step.

Fixes: #40

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>