Package extension even if icon is not present
If an extension specifies an icon in its package.json but the
specified file is not provided, 'vsce' executable will fail packaging

However it's quite common to have built-in extensions with no icons
so this fixes the issue by removing the icon property from the
package.json before building.

Once published the registry will apply a default icon.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>