Add external built-in extensions to extension packs
The source code of the built-in extensions is often maintained in the
same repository as the main application dependent on them, however
there is a small subset of these extensions where the source code
is maintained in a different repository, this change refer to them as
`external built-in extensions`.

This change hard codes a list of known `external built-in extensions`
and includes them within generated built-in extension packages.
These extensions correspond to the ones listed under vscode's
`product.json`.

The extension 'ms-vscode.vscode-js-profile-table' is excluded from
this change as it's not passing validation checks i.e. no publisher,
no version. In addition, it depends on additional extensions that use
a proposed API which is subject to change.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>