fix packaging issues on Windows
Tweak the scripts to allow building and packaging from Windows.

Use `execa` instead of `child_process` to spawn subprocesses as it
behaves better when invoking things like `yarn`.

Use `cross-env` to set environment variables in the npm scripts.

Update the types for `node` as we expect to run at least 14.

Cleanup some code.