temporarily disable publish to open-vsx
Fixes #93

1. disable open-vsx publish for solid revisions
Disable both nightly and when merging to master. This way we will not
further "pollute" the versions ahead of time, pass the current 1.62.3. We will
re-enable solid revision publish when we catch up and need more recent versions
published.

in the meantime we keep the build, so we see if problems arise.

2. disable cron nightly open-vsx publish for internediary revisions (next)

Keep publish when merging to master, so we can confirm that publishing still
works, but remove nightly CRON publishing.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>