[pack][js-debug] Fix ext pack issue that left-out js-debug*
We use node-fetch to call open-vsx API end-points. For external
built-in extensions, we want to make sure they exist on open-vsx
before adding them to the pack, to avoid issues when resolving the
pack(we do not check for a specific version, because they may not
follow vscode's versioning).

It turns-out that For this case, we were generating a URL that had
a trailing slash, which did not work:

Original URL, returns 404:
https://open-vsx.org/api/ms-vscode/js-debug-companion/

Corrected URL (works):
https://open-vsx.org/api/ms-vscode/js-debug-companion

P.S. also aligned version of `node-fetch` and its typings

Fixes #123

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>